+++ CHANGELOG.md
@@ -15,6 +15,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Issue **#1784** : Several functions were previously prevented from working on results from aggregate functions but are now applied regardless.
 
+* Fix config file validation not working when hot loading config file changes.
+
+* Change config file validation to be the first thing that happens on boot.
+
+* Fix error when empty branches are in the config file.
+
+* Add `pipeline.referenceData.getLmdbSystemLibraryPath` prop to support provided LMDB binary.
+
+* Change extraction location of bundled LMDB binary to be the same as the store files.
+
+* Change default value for `pipeline.referenceData.maxPutsBeforeCommit` to 0 (i.e. don't commit mid-load).
+
+
 ## [v7.0-beta.50] - 2020-09-07
 
 * Add /api/refData/v1/lookup REST endpoint for doing ref data lookups.